refactor commands; add struts for autocomplete
