+++ Cargo.toml
@@ -22,6 +22,14 @@ tokio = { version = "1.27.0", features = ["tracing", "macros", "rt-multi-thread"
 tracing = "0.1.37"
 tracing-subscriber = "0.3.16"
 
+[dependencies.meilisearch-sdk]
+version = "0.24.1"
+features = ["isahc-static-ssl"]
+
+[dependencies.serde]
+version= "1.0"
+features = ["derive"]
+
 [dependencies.rust-s3]
 version ="0.33.0"
 default-features = false
