+++ docker-compose.yml
@@ -12,4 +12,3 @@ services:
       billyjoule.env
     ports:
       - 59090:9090
-
